+++ site/app/Www/Presenters/CompanyTrainingsPresenter.php
@@ -6,7 +6,7 @@ namespace MichalSpacekCz\Www\Presenters;
 use MichalSpacekCz\Formatter\TexyFormatter;
 use MichalSpacekCz\Training\CompanyTrainings;
 use MichalSpacekCz\Training\Prices;
-use MichalSpacekCz\Training\Reviews;
+use MichalSpacekCz\Training\Reviews\TrainingReviews;
 use MichalSpacekCz\Training\TrainingLocales;
 use MichalSpacekCz\Training\Trainings;
 use Nette\Application\BadRequestException;
@@ -23,7 +23,7 @@ class CompanyTrainingsPresenter extends BasePresenter
 		private readonly Trainings $trainings,
 		private readonly CompanyTrainings $companyTrainings,
 		private readonly TrainingLocales $trainingLocales,
-		private readonly Reviews $trainingReviews,
+		private readonly TrainingReviews $trainingReviews,
 		private readonly Prices $prices,
 		private readonly IResponse $httpResponse,
 	) {
